Fix issue with storage query when there are similar rows in database. Add validation to report query resolver. Move SQL queries from report.ts to db_queries.ts.
#20
